+++ packages/iov-rise/src/riseconnection.spec.ts
@@ -3,7 +3,9 @@ import {
   Address,
   Amount,
   BcpAccountQuery,
+  BcpAddressQuery,
   BcpBlockInfo,
+  BcpPubkeyQuery,
   BcpTransactionState,
   SendTx,
   SignedTransaction,
@@ -159,12 +161,31 @@ describe("RiseConnection", () => {
 
   it("can get nonce", async () => {
     const connection = await RiseConnection.establish(base);
-    const query: BcpAccountQuery = { address: "5399275477602875017R" as Address };
-    const nonce = (await connection.getNonce(query)).data[0];
 
-    // nonce is current unix timestamp +/- one second
-    expect(nonce.toNumber()).toBeGreaterThanOrEqual(Date.now() / 1000 - 1);
-    expect(nonce.toNumber()).toBeLessThanOrEqual(Date.now() / 1000 + 1);
+    // by address
+    {
+      const query: BcpAddressQuery = { address: "5399275477602875017R" as Address };
+      const nonce = (await connection.getNonce(query)).data[0];
+      // nonce is current unix timestamp +/- one second
+      expect(nonce.toNumber()).toBeGreaterThanOrEqual(Date.now() / 1000 - 1);
+      expect(nonce.toNumber()).toBeLessThanOrEqual(Date.now() / 1000 + 1);
+    }
+
+    // by pubkey
+    {
+      const query: BcpPubkeyQuery = {
+        pubkey: {
+          algo: Algorithm.Ed25519,
+          data: fromHex("ac681190391fe048d133a60e9b49f7ac0a8b0500b58a9f176b88aee1e79fe735") as PublicKeyBytes,
+        },
+      };
+      const nonce = (await connection.getNonce(query)).data[0];
+      // nonce is current unix timestamp +/- one second
+      expect(nonce.toNumber()).toBeGreaterThanOrEqual(Date.now() / 1000 - 1);
+      expect(nonce.toNumber()).toBeLessThanOrEqual(Date.now() / 1000 + 1);
+    }
+
+    connection.disconnect();
   });
 
   describe("postTx", () => {